+++ os/include/compat/uapi/linux/in.h
@@ -6,7 +6,7 @@
 static inline uint16_t ntohs(uint16_t netshort)
 {
 #if  __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
-	return (netshort << 8) | (netshort >> 8);
+	return (uint16_t) (netshort << 8) | (uint16_t) (netshort >> 8);
 #else
 	return netshort;
 #endif
@@ -15,7 +15,7 @@ static inline uint16_t ntohs(uint16_t netshort)
 static inline uint16_t htons(uint16_t hostshort)
 {
 #if  __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
-	return (hostshort << 8) | (hostshort >> 8);
+	return (uint16_t) (hostshort << 8) | (uint16_t) (hostshort >> 8);
 #else
 	return hostshort;
 #endif
@@ -24,7 +24,7 @@ static inline uint16_t htons(uint16_t hostshort)
 static inline uint32_t htonl(uint32_t hostlong)
 {
 #if  __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
-	return (hostlong << 24) | ((hostlong << 8) & 0xFF) | ((hostlong >> 8) & 0xFF) | (hostlong >> 24);
+	return (uint32_t) (hostlong << 24) | (uint32_t) ((hostlong << 8) & 0xFF) | (uint32_t) ((hostlong >> 8) & 0xFF) | (uint32_t) (hostlong >> 24);
 #else
 	return hostlong;
 #endif
